+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/domain/Command.kt
@@ -55,6 +55,9 @@ enum class Command {
     FDDB,
     FDGB,
     ALIASDB,
+    ALIASBB,
+    CHARBB,
+    FDBB,
 
     //mizuumi
     ALIASMB,