+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/usecase/CreateEmbedUseCase.kt
@@ -100,7 +100,7 @@ internal class CreateEmbedUseCase {
             actionRow {
                 rowButtons.forEach { button ->
                     interactionButton(
-                        style = ButtonStyle.Secondary,
+                        style = ButtonStyle.Primary,
                         customId = button.query,
                     ) {
                         label = button.label