+++ core/src/commonMain/kotlin/io/github/sophon/core/wiki/domain/model/Character.kt
@@ -11,11 +11,14 @@ data class Character(
     val aliasList: List<String> = listOf(),
     val images: Images? = null,
 
+    val umo: List<String> = listOf(),
+
     val sf6Properties: SF6Properties? = null,
     val ggstProperties: GGSTProperties? = null,
     val mkProperties: MK1Properties? = null,
     val dbfzProperties: DBFZProperties? = null,
     val gbvsrProperties: GBVSRProperties? = null,
+    val bbProperties: BBProperties? = null,
 ) {
     @Serializable
     data class Images(
@@ -57,7 +60,6 @@ data class Character(
         val bwdDashAirborne: String?,
         val bwdDashDist: String?,
         val fwdDash: String?,
-        val umo: List<String> = listOf(),
         val jumpDuration: String?,
         val highJumpDuration: String?,
         val jumpHeight: String?,
@@ -84,14 +86,12 @@ data class Character(
     @Serializable
     data class DBFZProperties(
         val kiMod: String?,
-        val umo: String?,
     )
 
     @Serializable
     data class GBVSRProperties(
         val prejump: String?,
         val backdash: String?,
-        val umo: String?,
         val walkSpeed: String?,
         val walkSpeedRelative: String?,
         val walkSpeedBack: String?,
@@ -107,5 +107,13 @@ data class Character(
         val hp: String? = null,
         val hpMod: String? = null,
         val throwDmg: String? = null,
+    )
+
+    @Serializable
+    data class BBProperties(
+        val hp: String? = null,
+        val preJump: String? = null,
+        val backDash: String? = null,
+        val forwardDash: String? = null,
     )
 }