+++ feat/wikiDustLoop/src/commonMain/kotlin/io/github/sophon/wikidustloop/data/MoveListResponseDto.kt
@@ -37,6 +37,7 @@ internal data class MoveDto(
     val caption: String? = null,
     val hitboxCaption: String? = null,
 
+    //gg
     val riscGain: String? = null,
     val riscLoss: String? = null,
     val wallDamage: String? = null,
@@ -44,6 +45,7 @@ internal data class MoveDto(
     val chipRatio: String? = null,
     val OTGType: String? = null,
 
+    //db
     val attribute: String? = null,
     val smash: String? = null,
     val kigain: String? = null,
@@ -51,8 +53,21 @@ internal data class MoveDto(
     val groundHit: String? = null,
     val airHit: String? = null,
 
+    //gbvs
     val onCH: String? = null,
     val meter: String? = null,
     val cooldown: String? = null,
     val cls: String? = null,
+
+    //bb
+    val onODR: String? = null,
+    val p1: String? = null,
+    val p2: String? = null,
+    val starter: String? = null,
+    val groundCH: String? = null,
+    val airCH: String? = null,
+    val blockstop: String? = null,
+    val hitstop: String? = null,
+    val CHstop: String? = null,
+    val cancelTiming: String? = null,
 )